+++ cloudcmd/base_delete_cmd.go
@@ -52,3 +52,25 @@ func NewDeleteBaseCmd(factory internal.CloudFactory, deleteInterface DeleteInter
 
 	return deleteCmd
 }
+
+func shouldDelete(prompt internal.CloudPrompt, object string, f *pflag.FlagSet) (bool, error) {
+	confirmed, err := f.GetBool(confirm)
+	if err != nil {
+		return false, err
+	}
+
+	if confirmed {
+		return true, nil
+	}
+
+	return prompt.PerformActionPrompt(object)
+}
+
+type CloudObjectType int64
+
+const (
+	Org CloudObjectType = iota
+	Team
+	Env
+	Dep
+)
